bug 1079920 - Tokenize cells in compat table
Because of rowspan and colspan, we don't know if a cell is a feature or
a support until we walk the table.  This scraper change treats each cell
as a sequence of tokens - kuma functions, line breaks, code sections,
and plain text.  The tokens are then processed to remove whitespace,
apply flags, and identify canonically named features.  We're a step
closer to parsing supports and versions as well.